requirements.txt
django-scheduler==0.10.1